Add TSEM master header file.
TSEM is designed, from a functional perspective, to be contained
entirely in its own directory.

The tsem.h header file defines the enumeration types, structure
definitions and externally visiable functions that are referenced
by all of the compilation units of the TSEM LSM implementation in
that directory.

The structure and enumeration types are extensively documented
and are the recommended starting point for understanding TSEM
implementation and functionality.

Signed-off-by: Greg Wettstein <greg@enjellic.com>