Implement TSEM control plane.
The fs.c file contains the implementation of the TSEM control
plane that is rooted in the following directory in the securityfs
filesystem:

/sys/kernel/security/tsem

The following file documents the interface provided by the
control plane:

Documentation/ABI/testing/tsem

The pseudo-files act on the modeling context of the process that
is reading or writing to the control plane files.  For example,
reading the 'id' pseudo-file, returns the security modeling
namespace identifier that the process is running in.

The 'control' pseudo-file is the only writable file in the plane
and is used to control the TSEM implementation.  The most
important and primary roles are to create namespaces and set the
trust status of a process modeled by an external TMA.

The ExternalTMA directory is used to segregate the pseudo-files
that are created in order to surface security event descriptions
to an external trust orchestrator.  The files in this directory
appear as the numeric value of the modeling domain they were
created for.

The following directory:

/sys/kernel/security/tsem/InternalTMA

Holds directories that are used to surface the characteristics of
internal Trusted Modeling agents.  There is currently only a
single internal TMA implemented and its characteristics are in
the following directory:

/sys/kernel/security/tsem/InternalTMA/model0

This model is used to allow new TMA's to be implemented and to
maintain ABI compatibility if changes are needed in a TMA.

Signed-off-by: Greg Wettstein <greg@enjellic.com>